fix: move bitcoin related files to bitcoin package
