+++ packages/bitcoin/src/bitcoin-signer.ts
