+++ packages/bitcoin/src/native-segwit-account.hooks.ts
