+++ packages/bitcoin/src/taproot-account.hooks.ts
